requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
cffi==1.15.0
click==7.1.2
colorama==0.4.4
cryptography==36.0.1
Django==4.0.1
Flask==1.1.4
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
flask-restful-swagger==0.20.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.1
pycparser==2.21
PyMySQL==1.0.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
tzdata==2021.5
Werkzeug==1.0.1
